+++ lib/bolt/consumer/channel_delete.ex
@@ -1,7 +1,7 @@
 defmodule Bolt.Consumer.ChannelDelete do
   @moduledoc "Handles the `CHANNEL_DELETE` event."
 
-  alias Bolt.ModLog
+  alias Bolt.{ModLog, Starboard}
   alias Nostrum.Struct.Channel
 
   @spec handle(Channel.t()) :: nil | ModLog.on_emit()
@@ -15,11 +15,21 @@ defmodule Bolt.Consumer.ChannelDelete do
           _ -> "unknown channel type"
         end
 
-      ModLog.emit(
-        channel.guild_id,
-        "CHANNEL_DELETE",
-        "#{type_name} #{channel.name} (`#{channel.id}`) was deleted"
-      )
+      if Starboard.is_starboard_channel?(channel.guild_id, channel.id) do
+        Starboard.delete_data(channel.guild_id, channel.id)
+
+        ModLog.emit(
+          channel.guild_id,
+          "CHANNEL_DELETE",
+          "#{type_name} #{channel.name} (`#{channel.id}`) was deleted, and starboard configuration for the channel was deleted"
+        )
+      else
+        ModLog.emit(
+          channel.guild_id,
+          "CHANNEL_DELETE",
+          "#{type_name} #{channel.name} (`#{channel.id}`) was deleted"
+        )
+      end
     end
   end
 end