+++ lib/bolt/consumer/message_reaction_add.ex
@@ -1,11 +1,16 @@
 defmodule Bolt.Consumer.MessageReactionAdd do
   @moduledoc "Handles the `MESSAGE_REACTION_ADD` event."
 
-  alias Bolt.Paginator
-  alias Nostrum.Struct.Message.Reaction
+  alias Bolt.Paginatorolt.Starboard
+  alias Bolt.Starboard
+  alias Nostrum.Struct.Event.MessageReactionAdd
 
-  @spec handle(Reaction.t()) :: :ok
+  @spec handle(MessageReactionAdd.t()) :: :ok
   def handle(reaction) do
     GenServer.cast(Paginator, {:MESSAGE_REACTION_ADD, reaction})
+
+    if reaction.guild_id != nil and reaction.emoji.name == "⭐" do
+      Starboard.handle_star_reaction(reaction.guild_id, reaction.channel_id, reaction.message_id)
+    end
   end
 end